+++ src/app/loadout-builder/generated-sets/GeneratedSetButtons.tsx
@@ -1,3 +1,4 @@
+import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
 import { t } from 'app/i18next-t';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { applyLoadout } from 'app/loadout-drawer/loadout-apply';
@@ -20,6 +21,7 @@ export default function GeneratedSetButtons({
   store,
   set,
   notes,
+  params,
   canCompareLoadouts,
   halfTierMods,
   onLoadoutSet,
@@ -28,6 +30,7 @@ export default function GeneratedSetButtons({
   store: DimStore;
   set: ArmorSet;
   notes?: string;
+  params: LoadoutParameters;
   canCompareLoadouts: boolean;
   halfTierMods: PluggableInventoryItemDefinition[];
   onLoadoutSet(loadout: Loadout): void;
@@ -37,12 +40,12 @@ export default function GeneratedSetButtons({
 
   // Opens the loadout menu for the generated set
   const openLoadout = () => {
-    onLoadoutSet(createLoadout(store.classType, set, notes));
+    onLoadoutSet(createLoadout(store.classType, set, params, notes));
   };
 
   // Automatically equip items for this generated set to the active store
   const equipItems = () => {
-    const loadout = createLoadout(store.classType, set);
+    const loadout = createLoadout(store.classType, set, params, notes);
     return dispatch(applyLoadout(store, loadout, { allowUndo: true }));
   };
 
@@ -92,7 +95,12 @@ export default function GeneratedSetButtons({
 /**
  * Create a Loadout object, used for equipping or creating a new saved loadout
  */
-function createLoadout(classType: DestinyClass, set: ArmorSet, notes?: string): Loadout {
+function createLoadout(
+  classType: DestinyClass,
+  set: ArmorSet,
+  params: LoadoutParameters,
+  notes?: string
+): Loadout {
   const data = {
     tier: _.sumBy(Object.values(set.stats), statTier),
   };
@@ -102,5 +110,6 @@ function createLoadout(classType: DestinyClass, set: ArmorSet, notes?: string): 
   );
   loadout.classType = classType;
   loadout.notes = notes;
+  loadout.parameters = params;
   return loadout;
 }