+++ src/app/loadout-drawer/loadout-utils.ts
@@ -6,13 +6,15 @@ import { DimBucketType } from 'app/inventory/inventory-buckets';
 import { DimCharacterStat, DimStore } from 'app/inventory/store-types';
 import { SocketOverrides } from 'app/inventory/store/override-sockets';
 import { isPluggableItem } from 'app/inventory/store/sockets';
+import { isModStatActive } from 'app/loadout-builder/process/mappers';
 import { isLoadoutBuilderItem } from 'app/loadout/item-utils';
 import { isInsertableArmor2Mod, sortMods } from 'app/loadout/mod-utils';
 import { armorStats } from 'app/search/d2-known-values';
 import { emptyArray } from 'app/utils/empty';
 import { itemCanBeInLoadout } from 'app/utils/item-utils';
-import { DestinyClass, DestinyStatDefinition } from 'bungie-api-ts/destiny2';
-import { BucketHashes } from 'data/d2/generated-enums';
+import { getFirstSocketByCategoryHash } from 'app/utils/socket-utils';
+import { DestinyClass } from 'bungie-api-ts/destiny2';
+import { BucketHashes, SocketCategoryHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
 import { D2Categories } from '../destiny2/d2-bucket-categories';
@@ -140,28 +142,61 @@ export function getLight(store: DimStore, items: DimItem[]): number {
   }
 }
 
-/** Returns a map of armor stat hashes to stats. There should be just one of each item */
-export function getArmorStats(
-  defs: D1ManifestDefinitions | D2ManifestDefinitions,
-  items: DimItem[]
-): { [hash: number]: DimCharacterStat } {
-  const statDefs = armorStats.map((hash) => defs.Stat.get(hash) as DestinyStatDefinition);
+/**
+ * This gets the loadout stats for all the equipped items and mods.
+ *
+ * It will add all stats from the mods whether they are equipped or not. If
+ * you want to ensure it will be the same as the game stats, make sure to check
+ * if all mods will fit on the items.
+ */
+export function getLoadoutStats(
+  defs: D2ManifestDefinitions,
+  classType: DestinyClass,
+  subclass: LoadoutItem | undefined,
+  armor: DimItem[],
+  mods: PluggableInventoryItemDefinition[]
+) {
+  const statDefs = armorStats.map((hash) => defs.Stat.get(hash));
 
   // Construct map of stat hash to DimCharacterStat
-  const statsByArmorHash: { [hash: number]: DimCharacterStat } = {};
+  const stats: { [hash: number]: DimCharacterStat } = {};
   statDefs.forEach(({ hash, displayProperties: { description, icon, name } }) => {
-    statsByArmorHash[hash] = { hash, description, icon: bungieNetPath(icon), name, value: 0 };
+    stats[hash] = { hash, description, icon: bungieNetPath(icon), name, value: 0 };
   });
 
-  // Sum the items stats into the statsByArmorHash
-  items.forEach((item) => {
+  // Sum the items stats into the stats
+  armor.forEach((item) => {
     const itemStats = _.groupBy(item.stats, (stat) => stat.statHash);
-    Object.entries(statsByArmorHash).forEach(([hash, stat]) => {
-      stat.value += itemStats[hash]?.[0].value ?? 0;
+    const energySocket =
+      item.sockets && getFirstSocketByCategoryHash(item.sockets, SocketCategoryHashes.ArmorTier);
+    Object.entries(stats).forEach(([hash, stat]) => {
+      stat.value += itemStats[hash]?.[0].base ?? 0;
+      stat.value += energySocket?.plugged?.stats?.[hash] || 0;
     });
   });
 
-  return statsByArmorHash;
+  // Add stats that come from the subclass fragments
+  if (subclass?.socketOverrides) {
+    for (const plugHash of Object.values(subclass.socketOverrides)) {
+      const plug = defs.InventoryItem.get(plugHash);
+      for (const stat of plug.investmentStats) {
+        if (stat.statTypeHash in stats) {
+          stats[stat.statTypeHash].value += stat.value;
+        }
+      }
+    }
+  }
+
+  // Add the mod stats
+  for (const mod of mods) {
+    for (const stat of mod.investmentStats) {
+      if (stat.statTypeHash in stats && isModStatActive(classType, mod.hash, stat, mods)) {
+        stats[stat.statTypeHash].value += stat.value;
+      }
+    }
+  }
+
+  return stats;
 }
 
 // Generate an optimized item set (loadout items) based on a filtered set of items and a value function